dm-2800
add categories controller and helper removed module - so popular category logic can go in the helper class when that is implemented in the feature branch.
