+++ app/controllers/practices_controller.rb
@@ -491,17 +491,6 @@ class PracticesController < ApplicationController
     end
   end
 
-  def update_category_usage
-    query_val = params["query"]
-    chosen_categories = params["chosenCategories"]
-    cat_id = params["catId"]
-    if cat_id.present?
-      ahoy.track('Category selected', {category_id: cat_id})
-    else
-      store_chosen_categories(query_val, chosen_categories)
-    end
-  end
-
   def extend_editor_session_time
     if @current_session.present? && @current_session.user === current_user
       PracticeEditorSession.extend_current_session(@current_session)
