+++ config/routes.rb
@@ -10,7 +10,7 @@ Rails.application.routes.draw do
 
   get '/dashboard/export', action: 'export_metrics', controller: 'admin/dashboard', as: 'export_metrics'
   get '/session_time_remaining', action: 'session_time_remaining', controller: 'practices', as: 'session_time_remaining'
-  patch '/update_category_usage', action: 'update_category_usage', controller: 'practices', as: 'update_category_usage'
+  patch '/update_category_usage', action: 'update_category_usage', controller: 'categories', as: 'update_category_usage'
   patch '/extend_editor_session_time', action: 'extend_editor_session_time', controller: 'practices', as: 'extend_editor_session_time'
   patch '/close_edit_session', action: 'close_edit_session', controller: 'practices', as: 'close_edit_session'
   post '/accept_terms', action: 'accept_terms', controller: 'users', as: 'accept_terms'
